Initial setup for PR merge back to master.

+++ source/MaterialXGenShader/TypeDesc.h
@@ -27,7 +27,7 @@ using ChannelMap = std::unordered_map<char, int>;
 /// Descriptors for registered types can be retreived using TypeDesc::get(), see below.
 class TypeDesc
 {
-  public:
+public:
     enum BaseType
     {
         BASETYPE_NONE,
@@ -53,8 +53,9 @@ class TypeDesc
 
     /// Register a type descriptor for a MaterialX data type.
     /// Throws an exception if a type with the same name is already registered.
-    static const TypeDesc* registerType(const string& name, unsigned char basetype, unsigned char semantic = SEMANTIC_NONE,
-                                        size_t size = 1, bool editable = true, const ChannelMap& channelMapping = ChannelMap());
+    static const TypeDesc* registerType(const string& name, unsigned char basetype,
+        unsigned char semantic = SEMANTIC_NONE, int size = 1, bool editable = true,
+        const ChannelMap& channelMapping = ChannelMap());
 
     /// Get a type descriptor for given name.
     /// Throws an exception if no type with that name is found.
@@ -102,14 +103,15 @@ class TypeDesc
     /// Return true if the type is an aggregate of 4 floats.
     bool isFloat4() const { return _size == 4 && (_semantic == SEMANTIC_COLOR || _semantic == SEMANTIC_VECTOR); }
 
-  private:
-    TypeDesc(const string& name, unsigned char basetype, unsigned char semantic, size_t size,
-             bool editable, const ChannelMap& channelMapping);
+private:
+    TypeDesc(const string& name, unsigned char basetype,
+      unsigned char semantic, int size, bool editable,
+      const ChannelMap& channelMapping);
 
     const string _name;
     const unsigned char _basetype;
     const unsigned char _semantic;
-    const size_t _size;
+    const int _size;
     const bool _editable;
     const ChannelMap _channelMapping;
 };
